Updated specification to reflect Issue #1
Lots of great comments in Issue #1. I attempted to incorporate the
feedback in this commit. To stay consistent with other serialization
rules (for hashing) I pulled in the GS and RS character codes rather
than use the plus character, which could hypothetically appear in the
data to be encoded. I also used full STRING types for the data in
DOSE, which I'm not happy about as that could increase output size
considerably.