Update URI @ to .
A minor error which we caught recently is that the URI spec defines @ as
the delimiter between the signature and the keyId, but the QR
Alphanumeric spec does not support the @ character. This commit revises
the @ to be the . character instead, which means we don't have to change
the spec's rules around percent encoding fields only, and we don't have
to grow the URI by two bytes unnecessarily.