Add language discussing percent encoding details
There are two characters that the RFC for percent encoding URLs doesn't
require to be percent encoded, but which are unavailable in alphanumeric
QR. They are `~` and `_` and while I don't think they will be used
often, it's important to discuss what a developer should do when they
encounter those characters in a data field.

This change adds a table that specifies the printable lower 127
characters, which would be encoded by "regular" URI encoders, and which
don't appear in the Alphanumeric QR spec and thus should be encoded.

I took the direction of "only encode MORE than RFC requires, not less"
in an attempt to save developer time (potentially at the expense of a
few-bytes-larger output).  General rule is `RFC || QR` for whether to
encode, meaning that the only time we do not want to percent encode a
character is when the RFC does *not* say to encode AND the character is
present in the 45 codepoints of alphanumeric QR.